# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTM API Docs",
    "version": "1.0.1",
    "description": "CTM all endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/mint/": {
      "post": {
        "tags": [
          "User",
          "Label"
        ],
        "description": "User must be logged in to mint nft",
        "parameters": [
          {
            "name": "coverImage",
            "in": "formData",
            "required": true,
            "description": "Picture for music cover image(image/png, image/jpeg, image/jpg, image/webp, image/tiff)",
            "schema": {
              "type": "file"
            }
          },
          {
            "name": "music",
            "in": "formData",
            "required": true,
            "description": "Music file(audio/mp4, audio/mpeg, audio/opus, audio/mpeg4-generic, audio/ogg)",
            "schema": {
              "type": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully minted nft",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully minted nft."
                    },
                    "contractId": {
                      "type": "number",
                      "example": 68
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully minted nft."
                    },
                    "contractId": {
                      "type": "number",
                      "example": 68
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "quantity": {
                    "example": "any"
                  },
                  "genre": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}